@@ -5,7 +5,6 @@ import 'package:d_cart_admin/providers/login_provider.dart';
 import 'package:d_cart_admin/utils/mytheme.dart';
 import 'package:d_cart_admin/utils/responsive_builder.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter/rendering.dart';
 import 'package:go_router/go_router.dart';
 import 'package:provider/provider.dart';
 import 'package:rive/rive.dart';
@@ -26,6 +25,7 @@ class _LoginScreenState extends State<LoginScreen> {
   final TextEditingController _passController = TextEditingController();
   final FocusNode _emailFocus = FocusNode();
   final FocusNode _passFocus = FocusNode();
+  final GlobalKey<FormState> _formKey = GlobalKey();
 
   StateMachineController? machineController;
   SMIInput<bool>? isChecking;
@@ -35,21 +35,21 @@ class _LoginScreenState extends State<LoginScreen> {
   SMIInput<bool>? trigFail;
 
   late Size size;
-  late LoginProvider loginProvider;
+  // late LoginProvider loginProvider;
   @override
   void initState() {
-    loginProvider = Provider.of<LoginProvider>(context, listen: false);
+    // loginProvider = Provider.of<LoginProvider>(context, listen: false);
     _emailFocus.addListener(emailFocusListener);
     _passFocus.addListener(passFocusListener);
 
-    bool b = loginProvider.getLoggedInUser();
-    print("current user $b");
-    if (b) {
-      // Navigator.pushNamedAndRemoveUntil(context, MyRoute.dashboard, (route) => false);
-      WidgetsBinding.instance.addPostFrameCallback((_) {
+    // Navigator.pushNamedAndRemoveUntil(context, MyRoute.dashboard, (route) => false);
+    WidgetsBinding.instance.addPostFrameCallback((_) {
+      bool b = context.read<LoginProvider>().getLoggedInUser();
+      if (b) {
         context.go(MyRoute.dashboard);
-      });
-    }
+      }
+    });
+
     super.initState();
   }
 
@@ -105,132 +105,143 @@ class _LoginScreenState extends State<LoginScreen> {
         ),
       );
 
-  Widget get loginForm => Center(
-        child: Padding(
-          padding: const EdgeInsets.all(8.0),
-          child: Card(
-            shape: RoundedRectangleBorder(
-              borderRadius: BorderRadius.circular(10),
-            ),
-            elevation: 30,
-            child: Container(
-              padding: const EdgeInsets.all(20),
-              width: double.maxFinite,
-              child: SingleChildScrollView(
-                physics: const BouncingScrollPhysics(),
-                child: Column(
-                  mainAxisAlignment: MainAxisAlignment.center,
-                  crossAxisAlignment: CrossAxisAlignment.start,
-                  mainAxisSize: MainAxisSize.min,
-                  children: [
-                    const SizedBox(height: 10),
-                    const Text(
-                      "Signin",
-                      style: TextStyle(
-                        color: Colors.black,
-                        fontSize: 29,
-                        fontWeight: FontWeight.w500,
+  Widget get loginForm => Form(
+        key: _formKey,
+        child: Center(
+          child: Padding(
+            padding: const EdgeInsets.all(8.0),
+            child: Card(
+              shape: RoundedRectangleBorder(
+                borderRadius: BorderRadius.circular(10),
+              ),
+              elevation: 30,
+              child: Container(
+                padding: const EdgeInsets.all(20),
+                width: double.maxFinite,
+                child: SingleChildScrollView(
+                  physics: const BouncingScrollPhysics(),
+                  child: Column(
+                    mainAxisAlignment: MainAxisAlignment.center,
+                    crossAxisAlignment: CrossAxisAlignment.start,
+                    mainAxisSize: MainAxisSize.min,
+                    children: [
+                      const SizedBox(height: 10),
+                      const Text(
+                        "Signin",
+                        style: TextStyle(
+                          color: Colors.black,
+                          fontSize: 29,
+                          fontWeight: FontWeight.w500,
+                        ),
                       ),
-                    ),
-                    const Text(
-                      "Welcome back",
-                      style: TextStyle(
-                        color: Colors.grey,
-                        fontSize: 14,
+                      const Text(
+                        "Welcome back",
+                        style: TextStyle(
+                          color: Colors.grey,
+                          fontSize: 14,
+                        ),
                       ),
-                    ),
-                    Visibility(
-                      visible: true,
-                      child: Center(
-                        child: SizedBox(
-                          height: 200,
-                          width: 200,
-                          child: RiveAnimation.asset(
-                            "assets/rive_animation/login_anim.riv",
-                            stateMachines: const ["Login Machine"],
-                            // artboard: "assets/rive_animation/login_anim.riv",
-                            onInit: (board) {
-                              machineController = StateMachineController.fromArtboard(board, "Login Machine");
+                      Visibility(
+                        visible: true,
+                        child: Center(
+                          child: SizedBox(
+                            height: 200,
+                            width: 200,
+                            child: RiveAnimation.asset(
+                              "assets/rive_animation/login_anim.riv",
+                              stateMachines: const ["Login Machine"],
+                              // artboard: "assets/rive_animation/login_anim.riv",
+                              onInit: (board) {
+                                machineController = StateMachineController.fromArtboard(board, "Login Machine");
 
-                              if (machineController == null) {
-                                print("hi");
-                                return;
-                              }
+                                if (machineController == null) {
+                                  return;
+                                }
 
-                              board.addController(machineController!);
-                              isChecking = machineController?.findInput("isChecking");
-                              numLook = machineController?.findInput("numLook");
-                              isHandsUp = machineController?.findInput("isHandsUp");
-                              trigSuccess = machineController?.findInput("trigSuccess");
-                              trigFail = machineController?.findInput("trigFail");
-                            },
+                                board.addController(machineController!);
+                                isChecking = machineController?.findInput("isChecking");
+                                numLook = machineController?.findInput("numLook");
+                                isHandsUp = machineController?.findInput("isHandsUp");
+                                trigSuccess = machineController?.findInput("trigSuccess");
+                                trigFail = machineController?.findInput("trigFail");
+                              },
+                            ),
                           ),
                         ),
                       ),
-                    ),
-                    const SizedBox(height: 10),
-                    TextFormFieldComponent(
-                      title: "Your Email",
-                      focusNode: _emailFocus,
-                      hint: "email@address.com",
-                      controller: _emailController,
-                      onChange: (st) {
-                        numLook?.change(st.length.toDouble() * 2);
-                      },
-                    ),
-                    const SizedBox(height: 20),
-                    Consumer<LoginProvider>(
-                      builder: (_, ref, child) => TextFormFieldComponent(
-                        isObscure: ref.obscure,
+                      const SizedBox(height: 10),
+                      TextFormFieldComponent(
+                        title: "Your Email",
+                        focusNode: _emailFocus,
+                        hint: "email@address.com",
+                        controller: _emailController,
+                        validator: (st) {
+                          if (st?.isEmpty ?? true) {
+                            return 'Email is required';
+                          }
+                          return null;
+                        },
+                        onChange: (st) {
+                          numLook?.change(st.length.toDouble() * 2);
+                        },
+                      ),
+                      const SizedBox(height: 20),
+                      TextFormFieldComponent(
+                        isObscure: context.watch<LoginProvider>().obscure,
                         focusNode: _passFocus,
                         title: "Password",
                         controller: _passController,
                         hint: "6+ characters required",
+                        validator: (st) {
+                          if (st?.isEmpty ?? true) {
+                            return 'Password is required';
+                          }
+                          return null;
+                        },
                         suffixWidget: InkWell(
                           splashColor: Colors.white,
                           hoverColor: Colors.white,
                           onTap: () {
-                            ref.toggleObscure();
-                            isHandsUp?.change(ref.obscure);
+                            context.read<LoginProvider>().toggleObscure();
+                            isHandsUp?.change(context.read<LoginProvider>().obscure);
                           },
                           child: Icon(
-                            !ref.obscure ? Icons.visibility_off : Icons.visibility,
+                            !context.watch<LoginProvider>().obscure ? Icons.visibility_off : Icons.visibility,
                             color: MyTheme.textFormBorder,
                           ),
                         ),
                       ),
-                    ),
-                    const SizedBox(height: 10),
-                    Consumer<LoginProvider>(
-                      builder: (_, ref, child) => CheckboxListTile(
+                      const SizedBox(height: 10),
+                      CheckboxListTile(
                         contentPadding: EdgeInsets.zero,
                         dense: true,
                         checkboxShape: RoundedRectangleBorder(
                           borderRadius: BorderRadius.circular(3),
                         ),
-                        value: ref.checked,
+                        value: context.watch<LoginProvider>().checked,
                         checkColor: Colors.white,
                         activeColor: MyTheme.textFormBorder,
                         title: const Text("Remember me"),
                         controlAffinity: ListTileControlAffinity.leading,
                         visualDensity: const VisualDensity(horizontal: -4, vertical: -4),
                         onChanged: (checked) {
-                          ref.toggleCheck(checked ?? false);
+                          context.read<LoginProvider>().toggleCheck(checked ?? false);
                         },
                       ),
-                    ),
-                    const SizedBox(height: 10),
-                    Consumer<LoginProvider>(
-                      builder: (_, ref, child) => CommonButton(
-                        text: ref.isLoading ? "Loading..." : "Login",
+                      const SizedBox(height: 10),
+                      CommonButton(
+                        text: context.watch<LoginProvider>().isLoading ? "Loading..." : "Login",
                         textStyle: const TextStyle(
                           fontSize: 16,
                           fontWeight: FontWeight.w400,
                         ),
-                        onPressed: ref.isLoading
+                        onPressed: context.watch<LoginProvider>().isLoading
                             ? null
                             : () async {
-                                bool isLogin = await ref.login(_emailController.text.trim(), _passController.text);
+                                if (!(_formKey.currentState?.validate() ?? false)) {
+                                  return;
+                                }
+                                bool isLogin = await context.read<LoginProvider>().login(_emailController.text.trim(), _passController.text);
 
                                 trigFail?.change(!isLogin);
 
@@ -241,8 +252,8 @@ class _LoginScreenState extends State<LoginScreen> {
                         borderRadius: 8,
                         verticalPadding: 10,
                       ),
-                    ),
-                  ],
+                    ],
+                  ),
                 ),
               ),
             ),
